Fix preview transactions not loading in multi-account views and uncategorized splits not loading
