+++ packages/desktop-client/src/components/FinancesApp.tsx
@@ -14,7 +14,7 @@ import { BankSync } from './banksync';
 import { BankSyncStatus } from './BankSyncStatus';
 import { CommandBar } from './CommandBar';
 import { GlobalKeys } from './GlobalKeys';
-import { Category } from './mobile/budget/Category';
+import { CategoryPage } from './mobile/budget/CategoryPage';
 import { MobileNavTabs } from './mobile/MobileNavTabs';
 import { TransactionEdit } from './mobile/transactions/TransactionEdit';
 import { Notifications } from './Notifications';
@@ -296,7 +296,7 @@ export function FinancesApp() {
                   path="/categories/:id"
                   element={
                     <WideNotSupported>
-                      <Category />
+                      <CategoryPage />
                     </WideNotSupported>
                   }
                 />
