+++ packages/desktop-client/src/components/mobile/accounts/OnBudgetAccountTransactions.tsx
@@ -5,53 +5,40 @@ import { listen, send } from 'loot-core/platform/client/fetch';
 import { type Query } from 'loot-core/shared/query';
 import { isPreviewId } from 'loot-core/shared/transactions';
 import {
-  type AccountEntity,
+  type ScheduleEntity,
   type TransactionEntity,
 } from 'loot-core/types/models';
 
-import { syncAndDownload } from '@desktop-client/app/appSlice';
 import { TransactionListWithBalances } from '@desktop-client/components/mobile/transactions/TransactionListWithBalances';
-import { useAccountPreviewTransactions } from '@desktop-client/hooks/useAccountPreviewTransactions';
 import { SchedulesProvider } from '@desktop-client/hooks/useCachedSchedules';
 import { useDateFormat } from '@desktop-client/hooks/useDateFormat';
 import { useNavigate } from '@desktop-client/hooks/useNavigate';
+import { useOnBudgetAccounts } from '@desktop-client/hooks/useOnBudgetAccounts';
+import { usePreviewTransactions } from '@desktop-client/hooks/usePreviewTransactions';
 import { accountSchedulesQuery } from '@desktop-client/hooks/useSchedules';
 import { useTransactions } from '@desktop-client/hooks/useTransactions';
 import { useTransactionsSearch } from '@desktop-client/hooks/useTransactionsSearch';
 import { collapseModals, pushModal } from '@desktop-client/modals/modalsSlice';
 import * as queries from '@desktop-client/queries';
-import { markAccountRead } from '@desktop-client/queries/queriesSlice';
 import { useDispatch } from '@desktop-client/redux';
 import * as bindings from '@desktop-client/spreadsheet/bindings';
 
-export function AccountTransactions({
-  account,
-}: {
-  readonly account: AccountEntity;
-}) {
-  const schedulesQuery = useMemo(
-    () => accountSchedulesQuery(account.id),
-    [account.id],
-  );
+export function OnBudgetAccountTransactions() {
+  const schedulesQuery = useMemo(() => accountSchedulesQuery('onbudget'), []);
 
   return (
     <SchedulesProvider query={schedulesQuery}>
-      <TransactionListWithPreviews account={account} />
+      <TransactionListWithPreviews />
     </SchedulesProvider>
   );
 }
 
-function TransactionListWithPreviews({
-  account,
-}: {
-  readonly account: AccountEntity;
-}) {
+function TransactionListWithPreviews() {
   const { t } = useTranslation();
-
   const baseTransactionsQuery = useCallback(
     () =>
-      queries.transactions(account.id).options({ splits: 'all' }).select('*'),
-    [account.id],
+      queries.transactions('onbudget').options({ splits: 'all' }).select('*'),
+    [],
   );
 
   const [transactionsQuery, setTransactionsQuery] = useState<Query>(
@@ -59,34 +46,27 @@ function TransactionListWithPreviews({
   );
   const {
     transactions,
-    isLoading: isTransactionsLoading,
+    isLoading,
     reload: reloadTransactions,
     isLoadingMore,
     loadMore: loadMoreTransactions,
   } = useTransactions({
     query: transactionsQuery,
   });
+  const onBudgetAccounts = useOnBudgetAccounts();
+  const onBudgetAccountsFilter = useCallback(
+    (schedule: ScheduleEntity) =>
+      onBudgetAccounts.some(a => a.id === schedule._account),
+    [onBudgetAccounts],
+  );
 
-  const { previewTransactions, isLoading: isPreviewTransactionsLoading } =
-    useAccountPreviewTransactions({
-      accountId: account?.id,
-    });
+  const { previewTransactions } = usePreviewTransactions({
+    filter: onBudgetAccountsFilter,
+  });
 
   const dateFormat = useDateFormat() || 'MM/dd/yyyy';
   const dispatch = useDispatch();
   const navigate = useNavigate();
-
-  const onRefresh = useCallback(() => {
-    if (account.id) {
-      dispatch(syncAndDownload({ accountId: account.id }));
-    }
-  }, [account.id, dispatch]);
-
-  useEffect(() => {
-    if (account.id) {
-      dispatch(markAccountRead({ id: account.id }));
-    }
-  }, [account.id, dispatch]);
 
   useEffect(() => {
     return listen('sync-event', event => {
@@ -120,12 +100,9 @@ function TransactionListWithPreviews({
               name: 'scheduled-transaction-menu',
               options: {
                 transactionId: transaction.id,
-                onPost: async (transactionId, today = false) => {
+                onPost: async transactionId => {
                   const parts = transactionId.split('/');
-                  await send('schedule/post-transaction', {
-                    id: parts[1],
-                    today,
-                  });
+                  await send('schedule/post-transaction', { id: parts[1] });
                   dispatch(
                     collapseModals({
                       rootModalName: 'scheduled-transaction-menu',
@@ -161,16 +138,11 @@ function TransactionListWithPreviews({
     [dispatch, navigate],
   );
 
-  const balanceQueries = useMemo(
-    () =>
-      account
-        ? {
-            balance: bindings.accountBalance(account.id),
-            cleared: bindings.accountBalanceCleared(account.id),
-            uncleared: bindings.accountBalanceUncleared(account.id),
-          }
-        : { balance: bindings.allAccountBalance() },
-    [account],
+  const balanceBindings = useMemo(
+    () => ({
+      balance: bindings.onBudgetAccountBalance(),
+    }),
+    [],
   );
 
   const transactionsToDisplay = !isSearching
@@ -180,21 +152,15 @@ function TransactionListWithPreviews({
 
   return (
     <TransactionListWithBalances
-      isLoading={
-        isSearching ? isTransactionsLoading : isPreviewTransactionsLoading
-      }
+      isLoading={isLoading}
       transactions={transactionsToDisplay}
-      balance={balanceQueries.balance}
-      balanceCleared={balanceQueries.cleared}
-      balanceUncleared={balanceQueries.uncleared}
+      balance={balanceBindings.balance}
       isLoadingMore={isLoadingMore}
       onLoadMore={loadMoreTransactions}
-      searchPlaceholder={t('Search {{accountName}}', {
-        accountName: account.name,
-      })}
+      searchPlaceholder={t('Search On Budget Accounts')}
       onSearch={onSearch}
       onOpenTransaction={onOpenTransaction}
-      onRefresh={onRefresh}
+      showMakeTransfer={true}
     />
   );
 }
