+++ packages/desktop-client/src/components/mobile/budget/BudgetTable.jsx
@@ -1,21 +1,10 @@
 import React, { useCallback, useMemo } from 'react';
-import { GridList, GridListItem } from 'react-aria-components';
-import { Trans, useTranslation } from 'react-i18next';
+import { useTranslation } from 'react-i18next';
 
 import { Button } from '@actual-app/components/button';
-import { Card } from '@actual-app/components/card';
 import { useResponsive } from '@actual-app/components/hooks/useResponsive';
-import { SvgLogo } from '@actual-app/components/icons/logo';
-import {
-  SvgArrowThinLeft,
-  SvgArrowThinRight,
-  SvgCheveronRight,
-} from '@actual-app/components/icons/v1';
-import {
-  SvgArrowButtonDown1,
-  SvgCalendar,
-  SvgViewShow,
-} from '@actual-app/components/icons/v2';
+import { SvgCheveronRight } from '@actual-app/components/icons/v1';
+import { SvgViewShow } from '@actual-app/components/icons/v2';
 import { Label } from '@actual-app/components/label';
 import { styles } from '@actual-app/components/styles';
 import { Text } from '@actual-app/components/text';
@@ -25,32 +14,22 @@ import { AutoTextSize } from 'auto-text-size';
 
 import * as monthUtils from 'loot-core/shared/months';
 import { q } from 'loot-core/shared/query';
-import { groupById } from 'loot-core/shared/util';
 
 import { ExpenseGroupList } from './ExpenseGroupList';
 import { IncomeGroup } from './IncomeGroup';
 
 import { MOBILE_NAV_HEIGHT } from '@desktop-client/components/mobile/MobileNavTabs';
 import { PullToRefresh } from '@desktop-client/components/mobile/PullToRefresh';
-import { MobilePageHeader, Page } from '@desktop-client/components/Page';
 import { PrivacyFilter } from '@desktop-client/components/PrivacyFilter';
 import { CellValue } from '@desktop-client/components/spreadsheet/CellValue';
 import { SchedulesProvider } from '@desktop-client/hooks/useCachedSchedules';
-import { useCategories } from '@desktop-client/hooks/useCategories';
 import { useFormat } from '@desktop-client/hooks/useFormat';
-import { useLocale } from '@desktop-client/hooks/useLocale';
 import { useLocalPref } from '@desktop-client/hooks/useLocalPref';
-import { useNavigate } from '@desktop-client/hooks/useNavigate';
-import { useOverspentCategories } from '@desktop-client/hooks/useOverspentCategories';
 import { useSheetValue } from '@desktop-client/hooks/useSheetValue';
 import { useSyncedPref } from '@desktop-client/hooks/useSyncedPref';
-import { useUndo } from '@desktop-client/hooks/useUndo';
-import { pushModal } from '@desktop-client/modals/modalsSlice';
-import { useDispatch } from '@desktop-client/redux';
 import {
   envelopeBudget,
   trackingBudget,
-  uncategorizedCount,
 } from '@desktop-client/spreadsheet/bindings';
 
 export const ROW_HEIGHT = 50;
@@ -313,19 +292,12 @@ function BudgetGroups({
 export function BudgetTable({
   categoryGroups,
   month,
-  monthBounds,
-  onPrevMonth,
-  onNextMonth,
-  onCurrentMonth,
   onShowBudgetSummary,
   onBudgetAction,
   onRefresh,
   onEditCategoryGroup,
   onEditCategory,
-  onOpenBudgetPageMenu,
-  onOpenBudgetMonthMenu,
 }) {
-  const { t } = useTranslation();
   const { width } = useResponsive();
   const show3Columns = width >= 300;
 
@@ -348,54 +320,7 @@ export function BudgetTable({
   const schedulesQuery = useMemo(() => q('schedules').select('*'), []);
 
   return (
-    <Page
-      padding={0}
-      header={
-        <MobilePageHeader
-          title={
-            <MonthSelector
-              month={month}
-              monthBounds={monthBounds}
-              onOpenMonthMenu={onOpenBudgetMonthMenu}
-              onPrevMonth={onPrevMonth}
-              onNextMonth={onNextMonth}
-            />
-          }
-          leftContent={
-            <Button
-              variant="bare"
-              style={{ margin: 10 }}
-              onPress={onOpenBudgetPageMenu}
-              aria-label={t('Budget page menu')}
-            >
-              <SvgLogo
-                style={{ color: theme.mobileHeaderText }}
-                width="20"
-                height="20"
-              />
-              <SvgCheveronRight
-                style={{ flexShrink: 0, color: theme.mobileHeaderTextSubdued }}
-                width="14"
-                height="14"
-              />
-            </Button>
-          }
-          rightContent={
-            <Button
-              variant="bare"
-              onPress={onCurrentMonth}
-              aria-label={t('Today')}
-              style={{ margin: 10 }}
-            >
-              {!monthUtils.isCurrentMonth(month) && (
-                <SvgCalendar width={20} height={20} />
-              )}
-            </Button>
-          }
-        />
-      }
-    >
-      <Banners month={month} onBudgetAction={onBudgetAction} />
+    <>
       <BudgetTableHeader
         month={month}
         show3Columns={show3Columns}
@@ -426,305 +351,7 @@ export function BudgetTable({
           </SchedulesProvider>
         </View>
       </PullToRefresh>
-    </Page>
-  );
-}
-
-function Banner({ type = 'info', children }) {
-  return (
-    <Card
-      style={{
-        height: 50,
-        marginTop: 10,
-        marginBottom: 10,
-        padding: 10,
-        justifyContent: 'center',
-        backgroundColor:
-          type === 'critical'
-            ? theme.errorBackground
-            : type === 'warning'
-              ? theme.warningBackground
-              : theme.noticeBackground,
-      }}
-    >
-      {children}
-    </Card>
-  );
-}
-
-function UncategorizedTransactionsBanner(props) {
-  const count = useSheetValue(uncategorizedCount());
-  const navigate = useNavigate();
-
-  if (count === null || count <= 0) {
-    return null;
-  }
-
-  return (
-    <GridListItem textValue="Uncategorized transactions banner" {...props}>
-      <Banner type="warning">
-        <View
-          style={{
-            flex: 1,
-            flexDirection: 'row',
-            alignItems: 'center',
-            justifyContent: 'space-between',
-          }}
-        >
-          <Trans count={count}>
-            You have {{ count }} uncategorized transactions
-          </Trans>
-          <Button
-            onPress={() => navigate('/accounts/uncategorized')}
-            style={PILL_STYLE}
-          >
-            <Text>
-              <Trans>Categorize</Trans>
-            </Text>
-          </Button>
-        </View>
-      </Banner>
-    </GridListItem>
-  );
-}
-
-function OverbudgetedBanner({ month, onBudgetAction, ...props }) {
-  const { t } = useTranslation();
-  const toBudgetAmount = useSheetValue(envelopeBudget.toBudget);
-  const dispatch = useDispatch();
-  const { showUndoNotification } = useUndo();
-  const { list: categories } = useCategories();
-  const categoriesById = groupById(categories);
-
-  const openCoverOverbudgetedModal = useCallback(() => {
-    dispatch(
-      pushModal({
-        modal: {
-          name: 'cover',
-          options: {
-            title: t('Cover overbudgeted'),
-            month,
-            showToBeBudgeted: false,
-            onSubmit: categoryId => {
-              onBudgetAction(month, 'cover-overbudgeted', {
-                category: categoryId,
-              });
-              showUndoNotification({
-                message: t('Covered overbudgeted from {{categoryName}}', {
-                  categoryName: categoriesById[categoryId].name,
-                }),
-              });
-            },
-          },
-        },
-      }),
-    );
-  }, [
-    categoriesById,
-    dispatch,
-    month,
-    onBudgetAction,
-    showUndoNotification,
-    t,
-  ]);
-
-  if (!toBudgetAmount || toBudgetAmount >= 0) {
-    return null;
-  }
-
-  return (
-    <GridListItem textValue="Overbudgeted banner" {...props}>
-      <Banner type="critical">
-        <View
-          style={{
-            flexDirection: 'row',
-            alignItems: 'center',
-            justifyContent: 'space-between',
-          }}
-        >
-          <View
-            style={{
-              flexDirection: 'row',
-              alignItems: 'center',
-              gap: 10,
-            }}
-          >
-            <View
-              style={{
-                flexDirection: 'row',
-                alignItems: 'center',
-                gap: 10,
-              }}
-            >
-              <SvgArrowButtonDown1 style={{ width: 15, height: 15 }} />
-              <Text>
-                <Trans>You have budgeted more than your available funds</Trans>
-              </Text>
-            </View>
-          </View>
-          <Button onPress={openCoverOverbudgetedModal} style={PILL_STYLE}>
-            <Trans>Cover</Trans>
-          </Button>
-        </View>
-      </Banner>
-    </GridListItem>
-  );
-}
-
-function OverspendingBanner({ month, onBudgetAction, budgetType, ...props }) {
-  const { t } = useTranslation();
-
-  const { list: categories, grouped: categoryGroups } = useCategories();
-  const categoriesById = groupById(categories);
-  const groupsById = groupById(categoryGroups);
-
-  const dispatch = useDispatch();
-
-  const overspentCategories = useOverspentCategories({ month }).filter(c => {
-    if (budgetType === 'tracking') {
-      return !c.hidden && !groupsById[c.group].hidden;
-    }
-    return true;
-  });
-
-  const categoryGroupsToShow = useMemo(
-    () =>
-      categoryGroups
-        .filter(g => overspentCategories.some(c => c.group === g.id))
-        .map(g => ({
-          ...g,
-          categories: overspentCategories.filter(c => c.group === g.id),
-        })),
-    [categoryGroups, overspentCategories],
-  );
-
-  const { showUndoNotification } = useUndo();
-
-  const onOpenCoverCategoryModal = useCallback(
-    categoryId => {
-      const category = categoriesById[categoryId];
-      dispatch(
-        pushModal({
-          modal: {
-            name: 'cover',
-            options: {
-              title: category.name,
-              month,
-              categoryId: category.id,
-              onSubmit: fromCategoryId => {
-                onBudgetAction(month, 'cover-overspending', {
-                  to: category.id,
-                  from: fromCategoryId,
-                });
-                showUndoNotification({
-                  message: t(
-                    `Covered {{toCategoryName}} overspending from {{fromCategoryName}}.`,
-                    {
-                      toCategoryName: category.name,
-                      fromCategoryName:
-                        fromCategoryId === 'to-budget'
-                          ? 'To Budget'
-                          : categoriesById[fromCategoryId].name,
-                    },
-                  ),
-                });
-              },
-            },
-          },
-        }),
-      );
-    },
-    [categoriesById, dispatch, month, onBudgetAction, showUndoNotification, t],
-  );
-
-  const onOpenCategorySelectionModal = useCallback(() => {
-    dispatch(
-      pushModal({
-        modal: {
-          name: 'category-autocomplete',
-          options: {
-            title:
-              budgetType === 'envelope'
-                ? t('Cover overspending')
-                : t('Overspent categories'),
-            month,
-            categoryGroups: categoryGroupsToShow,
-            showHiddenCategories: true,
-            onSelect:
-              budgetType === 'envelope' ? onOpenCoverCategoryModal : null,
-            clearOnSelect: true,
-            closeOnSelect: false,
-          },
-        },
-      }),
-    );
-  }, [
-    categoryGroupsToShow,
-    dispatch,
-    month,
-    onOpenCoverCategoryModal,
-    t,
-    budgetType,
-  ]);
-
-  const numberOfOverspentCategories = overspentCategories.length;
-  if (numberOfOverspentCategories === 0) {
-    return null;
-  }
-
-  return (
-    <GridListItem textValue="Overspent banner" {...props}>
-      <Banner type="critical">
-        <View
-          style={{
-            flexDirection: 'row',
-            alignItems: 'center',
-            justifyContent: 'space-between',
-          }}
-        >
-          <View
-            style={{
-              flexDirection: 'row',
-              alignItems: 'center',
-              gap: 10,
-            }}
-          >
-            <Text>
-              <Trans count={numberOfOverspentCategories}>
-                You have {{ count: numberOfOverspentCategories }} overspent
-                categories
-              </Trans>
-            </Text>
-          </View>
-          <Button onPress={onOpenCategorySelectionModal} style={PILL_STYLE}>
-            {budgetType === 'envelope' && <Trans>Cover</Trans>}
-            {budgetType === 'tracking' && <Trans>View</Trans>}
-          </Button>
-        </View>
-      </Banner>
-    </GridListItem>
-  );
-}
-
-function Banners({ month, onBudgetAction }) {
-  const { t } = useTranslation();
-  const [budgetType = 'envelope'] = useSyncedPref('budgetType');
-
-  return (
-    <GridList
-      aria-label={t('Banners')}
-      style={{ backgroundColor: theme.mobilePageBackground }}
-    >
-      <UncategorizedTransactionsBanner />
-      <OverspendingBanner
-        month={month}
-        onBudgetAction={onBudgetAction}
-        budgetType={budgetType}
-      />
-      {budgetType === 'envelope' && (
-        <OverbudgetedBanner month={month} onBudgetAction={onBudgetAction} />
-      )}
-    </GridList>
+    </>
   );
 }
 
@@ -968,76 +595,6 @@ function BudgetTableHeader({
           )}
         </CellValue>
       </View>
-    </View>
-  );
-}
-
-function MonthSelector({
-  month,
-  monthBounds,
-  onOpenMonthMenu,
-  onPrevMonth,
-  onNextMonth,
-}) {
-  const locale = useLocale();
-  const { t } = useTranslation();
-  const prevEnabled = month > monthBounds.start;
-  const nextEnabled = month < monthUtils.subMonths(monthBounds.end, 1);
-
-  const arrowButtonStyle = {
-    padding: 10,
-    margin: 2,
-  };
-
-  return (
-    <View
-      style={{
-        flex: 1,
-        justifyContent: 'center',
-        alignItems: 'center',
-        flexDirection: 'row',
-      }}
-    >
-      <Button
-        aria-label={t('Previous month')}
-        variant="bare"
-        onPress={() => {
-          if (prevEnabled) {
-            onPrevMonth();
-          }
-        }}
-        style={{ ...arrowButtonStyle, opacity: prevEnabled ? 1 : 0.6 }}
-      >
-        <SvgArrowThinLeft width="15" height="15" />
-      </Button>
-      <Button
-        variant="bare"
-        style={{
-          textAlign: 'center',
-          fontSize: 16,
-          fontWeight: 500,
-        }}
-        onPress={() => {
-          onOpenMonthMenu?.(month);
-        }}
-        data-month={month}
-      >
-        <Text style={styles.underlinedText}>
-          {monthUtils.format(month, 'MMMM ‘yy', locale)}
-        </Text>
-      </Button>
-      <Button
-        aria-label={t('Next month')}
-        variant="bare"
-        onPress={() => {
-          if (nextEnabled) {
-            onNextMonth();
-          }
-        }}
-        style={{ ...arrowButtonStyle, opacity: nextEnabled ? 1 : 0.6 }}
-      >
-        <SvgArrowThinRight width="15" height="15" />
-      </Button>
     </View>
   );
 }
