+++ packages/desktop-client/src/components/mobile/transactions/TransactionList.tsx
@@ -33,10 +33,7 @@ import * as monthUtils from 'loot-core/shared/months';
 import { isPreviewId } from 'loot-core/shared/transactions';
 import { validForTransfer } from 'loot-core/shared/transfer';
 import { groupById, integerToCurrency } from 'loot-core/shared/util';
-import {
-  type AccountEntity,
-  type TransactionEntity,
-} from 'loot-core/types/models';
+import { type TransactionEntity } from 'loot-core/types/models';
 
 import { ROW_HEIGHT, TransactionListItem } from './TransactionListItem';
 
@@ -87,7 +84,7 @@ type TransactionListProps = {
   onOpenTransaction?: (transaction: TransactionEntity) => void;
   isLoadingMore: boolean;
   onLoadMore: () => void;
-  account?: AccountEntity;
+  showMakeTransfer?: boolean;
 };
 
 export function TransactionList({
@@ -96,7 +93,7 @@ export function TransactionList({
   onOpenTransaction,
   isLoadingMore,
   onLoadMore,
-  account,
+  showMakeTransfer = false,
 }: TransactionListProps) {
   const locale = useLocale();
   const { t } = useTranslation();
@@ -229,7 +226,7 @@ export function TransactionList({
       {selectedTransactions.size > 0 && (
         <SelectedTransactionsFloatingActionBar
           transactions={transactions}
-          showMakeTransfer={!account}
+          showMakeTransfer={showMakeTransfer}
         />
       )}
     </>
