+++ upcoming-release-notes/5072.md
@@ -1,0 +1,6 @@
+---
+category: Bugfix
+authors: [joel-jeremy]
+---
+
+Fix preview transactions not loading in multi-account views and uncategorized splits not loading
